feat:story1 AC4

+++ src/main/java/com/tw/apistackbase/repository/CaseRepository.java
@@ -3,5 +3,5 @@ package com.tw.apistackbase.repository;
 import com.tw.apistackbase.entity.Case;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CaseRepository extends JpaRepository<Case,Integer> {
+public interface CaseRepository extends JpaRepository<Case,Long> {
 }
